Create disk label and BAAS partition on first managament OS boot
There is a bug in the go-diskfs library that makes it impossible
to create EXT4 partitions. It might be nicer to switch to this on
the long term. The code snippet for that is:

```go
log.Warn("Create new machine partition and disk label")
theDisk, err := diskfs.Open(image.DeviceFile[:len(image.DeviceFile)-1],
diskfs.WithOpenMode(diskfs.ReadWrite))
if err != nil {
   log.Fatalf("Failed to open disk: %v", err)
}

// Create a GPT table with one partition to store the meta data for BAAS
table := &gpt.Table{
  LogicalSectorSize:  512,
  PhysicalSectorSize: 512,
  ProtectiveMBR:      false,
  Partitions: []*gpt.Partition{
    {
      Type: gpt.LinuxFilesystem,
      Start: 2 * 1024 * 1024,  // 2MB
      Size:  20 * 1024 * 1024, // 20MB
      Name: "BAASPartition",
    },
  },
 }

err = theDisk.Partition(table)
if err != nil {
  log.Fatalf("Failed to create partition: %v", err)
}

 _, err = theDisk.CreateFilesystem(disk.FilesystemSpec{
   Partition: 1,
   FSType:    filesystem.TypeExt4,
   VolumeLabel: "BAAS-fs",
})

if err != nil {
   log.Fatalf("Failed to create filesystem: %v", err)
}
```